added support for outputting FrSky telemetry by silpstream @ rcgroups feature TELEMETRY to enable it. when armed, telemetry is sent out the TX pin. It must be connected to 'Rx' pin of FrSky telemetry receiver via a level shifter. When disarmed, port goes back to 115200 bps for GUI. fixed a typo in imu.c introduced by  GYRO_INTERLEAVE junk. todo: just switch serial speed instead of reinitializing everything.
git-svn-id: https://afrodevices.googlecode.com/svn/trunk/baseflight@173 7c89a4a9-59b9-e629-4cfe-3a2d53b20e61

+++ src/gps.c
@@ -619,10 +619,10 @@ static uint32_t GPS_coord_to_degrees(char *s)
 }
 
 // helper functions
-static uint16_t grab_fields(char *src, uint8_t mult)
-{                               // convert string to uint16
+static uint32_t grab_fields(char *src, uint8_t mult)
+{                               // convert string to uint32
     uint8_t i;
-    uint16_t tmp = 0;
+    uint32_t tmp = 0;
     for (i = 0; src[i] != 0; i++) {
         if (src[i] == '.') {
             i++;
